My python script upto this point
This is the quick and temporary script I used for almost a year upto this point.

I want to pivot this to either Java or C++.

So what I would like to do is preen features and maintain good documenation.

Therefor, I will be able to code this thing in Java or C++ without a hitch, and possibly use it as a  first project for all subsequent programming languages I shall learn in future.

OKAY, BYE.